@@ -1,36 +1,33 @@
-[tool.poetry]
+[project]
 name = "todo-cli"
-version = "0.1.0"
-description = ""
-authors = ["Ajith Aravind <ajith.aravind100@gmail.com>"]
-license = "MIT"
+readme = "readme.md"
+license = { file = "license" }
+requires-python = ">=3.13"
+dynamic = ["version", "description", "authors", "urls", "keywords"]
+dependencies = ["pyyaml>=6.0.2,<7", "sqlalchemy>=2.0.36,<3"]
 
-[tool.poetry.dependencies]
-python = "^3.11"
-pyyaml = { version = "^6.0", platform = "linux" }
-sqlalchemy = { version = "^2.0.9", platform = "linux" }
-
-[tool.poetry.group.dev.dependencies]
-black = { version = "^23.3.0", platform = "linux" }
-pre-commit = { version = "^3.2.2", platform = "linux" }
-pytest = { version = "^7.2.2", platform = "linux" }
-pytest-cov = { version = "^4.0.0", platform = "linux" }
-ruff = { version = "^0.0.261", platform = "linux" }
-
-[build-system]
-requires = ["poetry-core>=1.0.0"]
-build-backend = "poetry.core.masonry.api"
+[project.optional-dependencies]
+dev = [
+    "pre-commit>=4.0.1,<5",
+    "pytest>=8.3.3,<9",
+    "pytest-cov>=6.0.0,<7",
+    "ruff>=0.7.3,<1",
+]
 
 [tool.ruff]
-select = ["A", "B", "C", "D", "E", "F", "I", "Q"]
-fixable = ["A", "B", "C", "D", "E", "F", "I", "Q"]
+target-version = "py313"
+line-length = 88
+indent-width = 4
+
+[tool.ruff.lint]
+select = ["A", "B", "C", "D", "E", "F", "I", "Q", "S", "W", "UP", "RUF"]
+fixable = ["A", "B", "C", "D", "E", "F", "I", "Q", "S", "W", "UP", "RUF"]
 ignore = ["D203", "D213"]
-line-length = 88
-target-version = "py311"
 
-[tool.ruff.per-file-ignores]
-"__init__.py" = ["F401", "D104"]
+[tool.ruff.lint.per-file-ignores]
+"__init__.py" = ["D104", "F401"]
+"tests/**/*.py" = ["S101"]
 
-[tool.black]
-line-length = 88
-target-version = ["py311"]
+[tool.ruff.format]
+quote-style = "double"
+indent-style = "space"
