@@ -10,7 +10,7 @@ CONFIG_FILE = "config.yaml"  # config file name
 # file path from parent dir
 CONFIG_FILE_PATH = Path(__file__).resolve().parent.parent.joinpath(CONFIG_FILE)
 
-with open(CONFIG_FILE_PATH, "r", encoding="utf8") as file:
+with open(CONFIG_FILE_PATH, encoding="utf8") as file:
     CONFIG = yaml.safe_load(file)
 
 
